Extract Excel XLSX package helpers
